Moved database related logic to the database folder
