+++ database/service/getOwnersAndFilters.js
@@ -1,0 +1,16 @@
+const db = require('../dbConfig');
+
+const getCarOwnersRecord = async () => {
+  const records = await db('car_owners');
+  return records;
+};
+
+const getFilters = async () => {
+  const filters = await db('car_filters');
+  return filters;
+};
+
+module.exports = {
+  getCarOwnersRecord,
+  getFilters,
+};
